requirements.txt
streamlit==1.32.0
supabase==2.3.4
pandas==2.2.1
numpy==1.26.4
plotly==5.19.0
python-dotenv==1.0.1
pydantic==2.6.3 